Create landing page
I've written all the html and css. Think I'm done, though it's not 100%
the same as the picture? It won't look the same if the screen is too large.
The header isn't linked up with the first second (the part that says this
website is awesome).

Some parts feel a bit "hacky." For example, I made the "Thor, God of Thunder"
part have a value of 20% for margin-right, which is probably not the
intended solution!